Refactor to a maven project
