+++ src/main/java/nschultz/filemanager/controller/AboutDialogController.java
@@ -1,0 +1,53 @@
+/*
+Copyright (c) 2017 Niklas Schultz
+Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
+rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+persons to whom the Software is furnished to do so, subject to the following conditions:
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
+Software. THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
+TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+DEALINGS IN THE SOFTWARE.
+ */
+package nschultz.filemanager.controller;
+
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Button;
+import javafx.scene.control.TextArea;
+import javafx.stage.Stage;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class AboutDialogController implements Initializable {
+
+  @FXML
+  private TextArea licenseArea;
+
+  @FXML
+  private Button closeButton;
+
+  @Override
+  public void initialize(URL location, ResourceBundle resources) {
+    licenseArea.setText("Copyright (c) 2017 Niklas Schultz\n"
+            + "Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated"
+            + "documentation files (the \"Software\"), to deal in the Software without restriction, including without limitation the"
+            + "rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit"
+            + "persons to whom the Software is furnished to do so, subject to the following conditions:\n"
+            + "The above copyright notice and this permission notice shall be included in all copies or substantial portions of the "
+            + "Software. THE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED"
+            + "TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. "
+            + "IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN"
+            + "ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER"
+            + "DEALINGS IN THE SOFTWARE.");
+  }
+
+  @FXML
+  private void closeButton_onAction(ActionEvent event) {
+    ((Stage) closeButton.getScene().getWindow()).close();
+  }
+}
