--- src/main/nschultz/filemanager/model/MainViewModel.java
@@ -1,69 +1,0 @@
-/*
-Copyright (c) 2017 Niklas Schultz
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
-rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit
-persons to whom the Software is furnished to do so, subject to the following conditions:
-The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
-Software. THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
-TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
-ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-DEALINGS IN THE SOFTWARE.
- */
-package main.nschultz.filemanager.model;
-
-import javafx.collections.ObservableList;
-import javafx.scene.input.Clipboard;
-import javafx.scene.input.ClipboardContent;
-import main.nschultz.filemanager.FileManagerApp;
-
-import java.io.File;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-
-public class MainViewModel {
-
-    public MainViewModel() {
-    }
-
-    public Path getPreviousDirFromFile(Path file) {
-        return file.getNameCount() == 1 ? file.getParent() : file.getParent().getParent();
-    }
-
-    public void openFileWithAssociatedProgram(Path filePath) {
-        FileManagerApp.fetchHostServices().showDocument(filePath.toString());
-    }
-
-    public boolean isFile(Path path) {
-        return Files.isRegularFile(path);
-    }
-
-    public List<String> getAccessibleDrives() {
-        List<String> drives = new ArrayList<>();
-        for (File drive : File.listRoots()) {
-            if (Files.isReadable(drive.toPath())) {
-                drives.add(drive.getAbsolutePath());
-            }
-        }
-        return drives;
-    }
-
-    public void copyStringToClipBoard(String stringToCopy) {
-        Clipboard clipboard = Clipboard.getSystemClipboard();
-        ClipboardContent clipboardContent = new ClipboardContent();
-        clipboardContent.putString(stringToCopy);
-        clipboard.setContent(clipboardContent);
-    }
-
-    public FileModel searchForFile(ObservableList<FileModel> list, String name) {
-        for (FileModel fileModel : list) {
-            if (fileModel.getName().toLowerCase().contains(name.toLowerCase())) {
-                return fileModel;
-            }
-        }
-        return null;
-    }
-}
