+++ src/main/resources/nschultz/filemanager/css/default.css
@@ -8,7 +8,7 @@
 
 }
 
-.table-row-cell: selected {
+.table-row-cell:selected {
     -fx-background-color: rgb(51, 174, 51);
 }
