+++ src/main/resources/nschultz/filemanager/fxml/MainView.fxml
@@ -20,7 +20,7 @@ DEALINGS IN THE SOFTWARE.
 
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="720"
             prefWidth="1420.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="main.nschultz.filemanager.controller.MainViewController">
+            fx:controller="nschultz.filemanager.controller.MainViewController">
     <children>
         <BorderPane prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
                     AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
